+++ changelog.md
@@ -8,6 +8,7 @@
 - Display badge for "now" when an event is happening now
 - Favorite events that have passed do not show in the list
 - Fix to correctly choose todays day if its available in events
+- Search now has a "go" button
 
 ### 1.1 - August 2, 2023
 - Submitted to the App Store